% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_lists.R
\name{check_input_lists}
\alias{check_input_lists}
\title{Check input lists for consistency}
\usage{
check_input_lists(student_list, lecturer_list, project_list)
}
\arguments{
\item{student_list}{list of student preferences}

\item{lecturer_list}{list of lecturer preferences and caps}

\item{project_list}{list of projects, lecturers, and caps}
}
\value{

}
\description{
Check input lists for consistency
}
