% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_lists.R
\name{check_lecturer_project_consistency}
\alias{check_lecturer_project_consistency}
\title{Check consistency of the lecturer and project inputs}
\usage{
check_lecturer_project_consistency(lecturer_list, project_list)
}
\arguments{
\item{lecturer_list}{list of lecturer preferences and caps}

\item{project_list}{list of projects, lecturers, and caps}
}
\value{

}
\description{
Check consistency of the lecturer and project inputs
}
