% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_lists.R
\name{check_project_student_consistency}
\alias{check_project_student_consistency}
\title{Check consistency of the project and student inputs}
\usage{
check_project_student_consistency(project_list, student_list)
}
\arguments{
\item{project_list}{list of projects, lecturers, and caps}

\item{student_list}{list of student preferences}
}
\value{

}
\description{
Check consistency of the project and student inputs
}
