% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_lists.R
\name{check_student_lecturer_consistency}
\alias{check_student_lecturer_consistency}
\title{Check consistency of the student and lecturer inputs}
\usage{
check_student_lecturer_consistency(student_list, lecturer_list)
}
\arguments{
\item{student_list}{list of student preferences}

\item{lecturer_list}{list of lecturer preferences and caps}
}
\value{

}
\description{
Check consistency of the student and lecturer inputs
}
