% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_capacity.R
\name{effective_capacity}
\alias{effective_capacity}
\title{Compute the effective capacity of lecturers}
\usage{
effective_capacity(lecturer_list, project_list)
}
\arguments{
\item{lecturer_list}{A list of lecturer preferences, in the format output by `read_lecturer_file`}

\item{project_list}{A list of project definitions, in the format output by `read_project_file`}
}
\value{

}
\description{
Compute the effective capacity of lecturers
}
