% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat_output_js.R
\name{neat_student_output}
\alias{neat_student_output}
\title{Create a neat (tibble) version of the student allocation}
\usage{
neat_student_output(allocation_output, student_list)
}
\arguments{
\item{allocation_output}{Output object from the allocation algorithm}

\item{student_list}{A list of student preferences, in the format output by `read_student_file`}
}
\value{

}
\description{
Create a neat (tibble) version of the student allocation
}
