% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projected_preferences.R
\name{projected_preference_list}
\alias{projected_preference_list}
\title{Initialize projected preferences and extend lecturer preferences}
\usage{
projected_preference_list(student_list, lecturer_list, project_list)
}
\arguments{
\item{student_list}{list of student preferences}

\item{lecturer_list}{list of lecturer preferences and caps}

\item{project_list}{list of projects, lecturers, and caps}
}
\value{

}
\description{
Initialize projected preferences and extend lecturer preferences
}
