% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_lecturer_file}
\alias{read_lecturer_file}
\title{Read in a lecturer preferences file}
\usage{
read_lecturer_file(filename)
}
\arguments{
\item{filename}{lecturer file name to read, with full path}
}
\value{

}
\description{
row i               : lecturer i's preferences, in descending order of preference \cr
rows have variable number of columns, each separated by whitespace \cr
column 1            : unique lecturer id (string) \cr
column 2            : lecturer capacity (integer) \cr
column 3-column n_i : list of student ids, in descending order of preference (strings) \cr
}
