% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_project_file}
\alias{read_project_file}
\title{Read in a project list file}
\usage{
read_project_file(filename)
}
\arguments{
\item{filename}{project file name to read, with full path}
}
\value{

}
\description{
row i               : project i specification 
rows have three columns, each separated by whitespace
column 1            : unique project id (string)
column 2            : project capacity (integer)
column 3            : lecturer that offers project (string)
}
