% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_student_file}
\alias{read_student_file}
\title{Read in a student project preferences file}
\usage{
read_student_file(filename)
}
\arguments{
\item{filename}{student file name to read, with full path}
}
\value{

}
\description{
row i               : student i's preferred projects, in descending order of preference \cr 
Rows have variable number of columns, each separated by whitespace \cr 
column 1            : unique student id (string)  \cr 
column 2-column n_i : list of project names (strings)  \cr
}
