% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spa_student.R
\name{shiny_spa_student}
\alias{shiny_spa_student}
\title{Run the shiny app for the SPA student algorithm}
\usage{
shiny_spa_student()
}
\value{

}
\description{
Run the shiny app for the SPA student algorithm
}
