% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat_output_js.R
\name{neat_project_output}
\alias{neat_project_output}
\title{Create a neat (tibble) version of the project allocation}
\usage{
neat_project_output(allocation_output, delim = pkg_options()$neat_delim)
}
\arguments{
\item{allocation_output}{Output object from the allocation algorithm}

\item{delim}{Character used to paste together multiple elements in a field in neat output}
}
\value{

}
\description{
Create a neat (tibble) version of the project allocation
}
